+++ src/editor/core/command/CommandAdapt.ts
@@ -16,12 +16,13 @@ import { IAppendElementListOption, IDrawImagePayload, IGetValueOption, IPainterO
 import { IEditorData, IEditorOption, IEditorResult } from '../../interface/Editor'
 import { IElement, IElementStyle } from '../../interface/Element'
 import { IMargin } from '../../interface/Margin'
+import { RangeContext } from '../../interface/Range'
 import { IColgroup } from '../../interface/table/Colgroup'
 import { ITd } from '../../interface/table/Td'
 import { ITr } from '../../interface/table/Tr'
 import { IWatermark } from '../../interface/Watermark'
-import { downloadFile, getUUID } from '../../utils'
-import { formatElementContext, formatElementList, isTextLikeElement } from '../../utils/element'
+import { deepClone, downloadFile, getUUID } from '../../utils'
+import { formatElementContext, formatElementList, isTextLikeElement, pickElementAttr } from '../../utils/element'
 import { printImageBase64 } from '../../utils/print'
 import { Control } from '../draw/control/Control'
 import { Draw } from '../draw/Draw'
@@ -34,10 +35,7 @@ import { Position } from '../position/Position'
 import { RangeManager } from '../range/RangeManager'
 import { WorkerManager } from '../worker/WorkerManager'
 
-
 export class CommandAdapt {
-
-  private readonly defaultWidth: number = 40
 
   private draw: Draw
   private range: RangeManager
@@ -532,7 +530,7 @@ export class CommandAdapt {
     for (let r = 0; r < row; r++) {
       const tdList: ITd[] = []
       const tr: ITr = {
-        height: 40,
+        height: this.options.defaultTrMinHeight,
         tdList
       }
       for (let c = 0; c < col; c++) {
@@ -725,7 +723,7 @@ export class CommandAdapt {
     // 重新计算宽度
     const colgroup = element.colgroup!
     colgroup.splice(curTdIndex, 0, {
-      width: this.defaultWidth
+      width: this.options.defaultColMinWidth
     })
     const colgroupWidth = colgroup.reduce((pre, cur) => pre + cur.width, 0)
     const width = this.draw.getOriginalInnerWidth()
@@ -782,7 +780,7 @@ export class CommandAdapt {
     // 重新计算宽度
     const colgroup = element.colgroup!
     colgroup.splice(curTdIndex, 0, {
-      width: this.defaultWidth
+      width: this.options.defaultColMinWidth
     })
     const colgroupWidth = colgroup.reduce((pre, cur) => pre + cur.width, 0)
     const width = this.draw.getOriginalInnerWidth()
@@ -1599,6 +1597,29 @@ export class CommandAdapt {
 
   public getRangeText(): string {
     return this.range.toString()
+  }
+
+  public getRangeContext(): RangeContext | null {
+    const range = this.range.getRange()
+    const { startIndex, endIndex } = range
+    if (!~startIndex && !~endIndex) return null
+    // 选区信息
+    const isCollapsed = startIndex === endIndex
+    // 元素信息
+    const elementList = this.draw.getElementList()
+    const startElement = pickElementAttr(elementList[isCollapsed ? startIndex : startIndex + 1])
+    const endElement = pickElementAttr(elementList[endIndex])
+    // 页码信息
+    const positionList = this.position.getPositionList()
+    const startPageNo = positionList[startIndex].pageNo
+    const endPageNo = positionList[endIndex].pageNo
+    return deepClone({
+      isCollapsed,
+      startElement,
+      endElement,
+      startPageNo,
+      endPageNo
+    })
   }
 
   public pageMode(payload: PageMode) {