+++ src/editor/core/contextmenu/menus/controlMenus.ts
@@ -6,7 +6,7 @@ export const controlMenus: IRegisterContextMenu[] = [
   {
     i18nPath: 'contextmenu.control.delete',
     when: (payload) => {
-      return !payload.editorHasSelection && payload.startElement?.type === ElementType.CONTROL
+      return !payload.isReadonly && !payload.editorHasSelection && payload.startElement?.type === ElementType.CONTROL
     },
     callback: (command: Command) => {
       command.executeRemoveControl()