+++ src/editor/interface/Range.ts
@@ -1,4 +1,5 @@
 import { EditorZone } from '../dataset/enum/Editor'
+import { IElement } from './Element'
 
 export interface IRange {
   startIndex: number;
@@ -15,3 +16,11 @@ export interface IRange {
 export type RangeRowArray = Map<number, number[]>
 
 export type RangeRowMap = Map<number, Set<number>>
+
+export type RangeContext = {
+  isCollapsed: boolean;
+  startElement: IElement;
+  endElement: IElement;
+  startPageNo: number;
+  endPageNo: number;
+}